backend/requirements/common.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==3.8.1
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-auth-ldap==4.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fonttools==4.31.2
future==0.18.2
flower==1.0.0
idna==3.3
kiwisolver==1.4.2
kombu==5.2.4
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-ldap==3.4.0
python3-openid==3.2.0
pytz==2021.3
redis==4.2.2
rep-api==0.3.27
requests==2.11.1
requests-oauthlib==1.3.1
simplejson==3.17.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
gunicorn==20.1.0
django-polymorphic==3.1.0